+++ src/s2ir.c
@@ -867,7 +867,7 @@ void LabelStatement::toIR(IRState *irs)
          */
         if (fwdrefs)
         {
-            for (int i = 0; i < fwdrefs->dim; i++)
+            for (size_t i = 0; i < fwdrefs->dim; i++)
             {   block *b = fwdrefs->tdata()[i];
 
                 if (b->Btry != lblock->Btry)
@@ -1520,7 +1520,7 @@ void TryCatchStatement::toIR(IRState *irs)
     block_next(blx,BCgoto,NULL);
 
     assert(catches);
-    for (int i = 0 ; i < catches->dim; i++)
+    for (size_t i = 0 ; i < catches->dim; i++)
     {
         Catch *cs = catches->tdata()[i];
         if (cs->var)
